CORE-12500 Combine TestGroupPolicyProviders & write integration test for new methods of utxoPersistenceService
