+++ components/membership/registration-impl/build.gradle
@@ -50,6 +50,7 @@ dependencies {
     testRuntimeOnly "org.slf4j:slf4j-simple:$slf4jVersion"
 
     integrationTestImplementation project(':testing:db-message-bus-testkit')
+    integrationTestImplementation project(":components:membership:group-policy-test-common")
 
     integrationTestRuntimeOnly project(':components:configuration:configuration-read-service-impl')
     integrationTestRuntimeOnly project(':components:crypto:crypto-client-hsm-impl')
