+++ settings.gradle
@@ -132,6 +132,7 @@ include 'components:membership:group-policy'
 include 'components:membership:group-policy-impl'
 include 'components:membership:group-policy-configuration-validation'
 include 'components:membership:group-policy-configuration-validation-impl'
+include 'components:membership:group-policy-test-common'
 include 'components:membership:membership-client'
 include 'components:membership:membership-client-impl'
 include 'components:membership:membership-persistence-client'
